split managers
