+++ alphadia/outputtransform/df_builders.py
@@ -7,8 +7,10 @@ import pandas as pd
 from alphabase.spectral_library.base import SpecLibBase
 
 from alphadia.constants.keys import StatOutputKeys
-from alphadia.workflow import manager
+from alphadia.workflow.managers.calibration_manager import CalibrationManager
+from alphadia.workflow.managers.optimization_manager import OptimizationManager
 from alphadia.workflow.managers.raw_file_manager import RawFileManager
+from alphadia.workflow.managers.timing_manager import TimingManager
 from alphadia.workflow.peptidecentric.peptidecentric import PeptideCentricWorkflow
 
 logger = logging.getLogger()
@@ -74,9 +76,7 @@ def build_run_stat_df(
                 PeptideCentricWorkflow.OPTIMIZATION_MANAGER_PKL_NAME,
             )
         ):
-            optimization_manager = manager.OptimizationManager(
-                path=optimization_manager_path
-            )
+            optimization_manager = OptimizationManager(path=optimization_manager_path)
             optimization_stats[StatOutputKeys.MS2_ERROR] = (
                 optimization_manager.ms2_error
             )
@@ -108,9 +108,7 @@ def build_run_stat_df(
                 PeptideCentricWorkflow.CALIBRATION_MANAGER_PKL_NAME,
             )
         ):
-            calibration_manager = manager.CalibrationManager(
-                path=calibration_manager_path
-            )
+            calibration_manager = CalibrationManager(path=calibration_manager_path)
 
             if (
                 fragment_mz_estimator := calibration_manager.get_estimator(
@@ -211,7 +209,7 @@ def build_run_internal_df(
     }
 
     if os.path.exists(timing_manager_path):
-        timing_manager = manager.TimingManager(path=timing_manager_path)
+        timing_manager = TimingManager(path=timing_manager_path)
         for key in timing_manager.timings:
             internal_dict[f"duration_{key}"] = [timing_manager.timings[key]["duration"]]
