+++ tests/unit_tests/test_outputtransform.py
@@ -7,8 +7,9 @@ import pandas as pd
 from conftest import mock_fragment_df, mock_precursor_df
 
 from alphadia.outputtransform.search_plan_output import SearchPlanOutput
-from alphadia.workflow import manager
 from alphadia.workflow.base import QUANT_FOLDER_NAME
+from alphadia.workflow.managers.optimization_manager import OptimizationManager
+from alphadia.workflow.managers.timing_manager import TimingManager
 from alphadia.workflow.peptidecentric.peptidecentric import PeptideCentricWorkflow
 
 
@@ -76,7 +77,7 @@ def test_output_transform():
         frag_df.to_parquet(os.path.join(raw_folder, "frag.parquet"), index=False)
         psm_df.to_parquet(os.path.join(raw_folder, "psm.parquet"), index=False)
 
-        optimization_manager = manager.OptimizationManager(
+        optimization_manager = OptimizationManager(
             config,
             path=os.path.join(
                 raw_folder,
@@ -84,7 +85,7 @@ def test_output_transform():
             ),
         )
 
-        timing_manager = manager.TimingManager(
+        timing_manager = TimingManager(
             path=os.path.join(
                 raw_folder,
                 PeptideCentricWorkflow.TIMING_MANAGER_PKL_NAME,
